Move component into page
