--- lib/components/__tests__/select-region.js
@@ -1,6 +1,0 @@
-import {testAndSnapshot} from 'lib/utils/component'
-import {mockRegion} from 'lib/utils/mock-data'
-
-import SelectRegion from '../select-region'
-
-testAndSnapshot(SelectRegion, {regions: [mockRegion]})
